rename: plugin_magefile -> pluginmage
