+++ magefiles/binaries.go
@@ -1,10 +1,10 @@
 package main
 
-import "github.com/mattermost/mattermost-plugin-legal-hold/plugin_magefile"
+import "github.com/mattermost/mattermost-plugin-legal-hold/pluginmage"
 
 func init() {
 	// Register the processor binary
-	plugin_magefile.RegisterBinary(plugin_magefile.BinaryBuildConfig{
+	pluginmage.RegisterBinary(pluginmage.BinaryBuildConfig{
 		Name:             "processor",
 		OutputPath:       "./bin",
 		BinaryNameFormat: "processor-{{.Manifest.Version}}-{{.GOOS}}-{{.GOARCH}}",
