+++ pluginmage/assets/.editorconfig
