+++ pluginmage/assets/.github/workflows/ci.yml
