+++ pluginmage/assets/.golangci.yml
